Add name from selection
During reviewing, a person can select a name and right click and a menu item item will appear that when clicked will save the name to the names.txt file . Future things we can add the functionallity to the browser not sure how many people would have a use case for that but it won't hurt. And add a shortcut. I didn't add a shortcut(even though I really need one) is because it might ruin the user experience having too many shortcuts. Wanted to hear your thoughts on this